requirements.txt
scipy>=1.7.1
opencv-python>=4.5.4
pillow>=8.4.0
pandas>=1.3.4
pytorch-lightning>=1.6.0
omegaconf>=2.0.0
einops>=0.3.2
tokenizers>=0.10.3
sklearn>=1.0.1
axial-positional-embedding
wandb>=0.12.5